backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.14.1
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
email-validator==2.2.0
python-dotenv==1.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pypdf==5.3.1
httpx==0.28.1
langgraph==0.2.66
pytest==8.3.5
